+++ packages/dumb-worker-router/src/fetcher.test.ts
@@ -98,4 +98,35 @@ describe("Router", () => {
     expect(response.status).toBe(200);
     expect(await response.json()).toBe("foobar");
   });
+
+  test("typed get fetch several paths", async () => {
+    const router = Router()
+      .get("/a/:param1/:param2", ({ params: { param1, param2 } }) => {
+        return ok(200, param1 + param2);
+      })
+      .get("/a/:param1", ({ params: { param1 } }) => {
+        return ok(200, param1);
+      });
+
+    const fetchMock = async (url: string, init?: RequestInit) => {
+      return router.handle(new Request(url, init));
+    };
+
+    const f = fetcher<RoutesOf<typeof router>>(fetchMock, {
+      origin: "http://t.co",
+    });
+
+    const response1 = await f.get("/a/:param1/:param2", {
+      params: { param1: "foo", param2: "bar" },
+    });
+
+    const response2 = await f.get("/a/:param1", {
+      params: { param1: "foo" },
+    });
+
+    expect(response1.status).toBe(200);
+    expect(await response1.json()).toBe("foobar");
+    expect(response2.status).toBe(200);
+    expect(await response2.json()).toBe("foo");
+  });
 });