+++ packages/dumb-worker-router/CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ...
+
+## [0.0.6] - 2023-04-17
+
+### Added
+
+- Added `*` at end
+- Added legible type errors on pattern issues for methods
+
+### Fixed
+
+- Fixed overlapping types
+
 ## [0.0.5] - 2023-04-17
 
 ### Added